refactor: remove "Next" from new component names
