+++ operate/client/src/App/DecisionInstance/Header/index.tsx
@@ -8,8 +8,8 @@
 
 import {Button} from '@carbon/react';
 import {tracking} from 'modules/tracking';
-import {InstanceHeader} from 'modules/components/InstanceHeaderNext';
-import {Skeleton} from 'modules/components/InstanceHeaderNext/Skeleton';
+import {InstanceHeader} from 'modules/components/InstanceHeader';
+import {Skeleton} from 'modules/components/InstanceHeader/Skeleton';
 import {Link} from 'modules/components/Link';
 import {Locations, Paths} from 'modules/Routes';
 import {formatDate} from 'modules/utils/date';
