+++ operate/client/src/App/ProcessInstance/ProcessInstanceHeader/index.tsx
@@ -14,8 +14,8 @@ import {Link} from 'modules/components/Link';
 import {Locations} from 'modules/Routes';
 import {panelStatesStore} from 'modules/stores/panelStates';
 import {tracking} from 'modules/tracking';
-import {InstanceHeader} from 'modules/components/InstanceHeaderNext';
-import {Skeleton} from 'modules/components/InstanceHeaderNext/Skeleton';
+import {InstanceHeader} from 'modules/components/InstanceHeader';
+import {Skeleton} from 'modules/components/InstanceHeader/Skeleton';
 import {VersionTag} from './styled';
 import {useProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
 import {useProcessInstanceXml} from 'modules/queries/processDefinitions/useProcessInstanceXml';
