+++ operate/client/src/modules/components/Diagram/DiagramControls/index.tsx
