+++ operate/client/src/modules/components/Diagram/DiagramControls/styled.ts
